Checking in VCDiff algorithm implementation that turns out to be completely un-useful...
But I worked so hard on it! I want to keep it for now :)

+++ Inversion.Core/Data/GitPackIndexEntry.cs
@@ -7,10 +7,10 @@ namespace Inversion.Data
 {
     public class GitPackIndexEntry
     {
-        public uint Offset { get; private set; }
+        public long Offset { get; private set; }
         public byte[] Hash { get; private set; }
 
-        public GitPackIndexEntry(uint offset, byte[] hash)
+        public GitPackIndexEntry(long offset, byte[] hash)
         {
             Offset = offset;
             Hash = hash;
